+++ src/builder.rs
@@ -1,9 +1,8 @@
 use crate::alloc::sync::Arc;
 use crate::callback::SessionCallback;
-use crate::config::{SASLConfig, SorterFn};
+use crate::config::SASLConfig;
 use crate::error::SASLError;
-use crate::registry::{Mechanism, Registry};
-use core::cmp::Ordering;
+use crate::registry::Registry;
 use core::fmt::{Debug, Formatter};
 
 #[derive(Clone)]
@@ -28,7 +27,6 @@ use core::fmt::{Debug, Formatter};
 /// use rsasl::config::SASLConfig;
 /// let config: Arc<SASLConfig> = SASLConfig::builder()
 ///     .with_default_mechanisms()
-///     .with_defaults()
 ///     .with_callback(Callback::new())
 ///     .unwrap();
 /// # }
@@ -72,7 +70,6 @@ use core::fmt::{Debug, Formatter};
 /// static MECHANISMS: &[Mechanism] = &[PLAIN, EXTERNAL];
 /// let config: Arc<SASLConfig> = SASLConfig::builder()
 ///     .with_registry(Registry::with_mechanisms(MECHANISMS))
-///     .with_defaults()
 ///     .with_callback(Callback::new())
 ///     .unwrap();
 /// ```
@@ -86,10 +83,6 @@ impl<State: Debug> Debug for ConfigBuilder<State> {
             .field("state", &self.state)
             .finish()
     }
-}
-
-pub fn default_sorter(a: &Mechanism, b: &Mechanism) -> Ordering {
-    a.priority.cmp(&b.priority)
 }
 
 #[derive(Clone, Debug)]
@@ -111,16 +104,15 @@ impl ConfigBuilder {
         ConfigBuilder {
             state: WantCallback {
                 mechanisms: Registry::default(),
-                sorter: default_sorter,
             },
         }
     }
 
     /// Use a pre-initialized mechanism registry, giving the most control over available mechanisms
     #[must_use]
-    pub const fn with_registry(self, mechanisms: Registry) -> ConfigBuilder<WantSorter> {
+    pub const fn with_registry(self, mechanisms: Registry) -> ConfigBuilder<WantCallback> {
         ConfigBuilder {
-            state: WantSorter { mechanisms },
+            state: WantCallback { mechanisms },
         }
     }
 
@@ -129,32 +121,12 @@ impl ConfigBuilder {
     /// This is equivalent to `Self::with_registry(Registry::default())`. The default set of
     /// mechanisms depends on the enabled cargo features.
     #[must_use]
-    pub fn with_default_mechanisms(self) -> ConfigBuilder<WantSorter> {
+    pub fn with_default_mechanisms(self) -> ConfigBuilder<WantCallback> {
         self.with_registry(Registry::default())
     }
 
-    pub(crate) fn with_credentials_mechanisms(self, authzid: bool) -> ConfigBuilder<WantSorter> {
+    pub(crate) fn with_credentials_mechanisms(self, authzid: bool) -> ConfigBuilder<WantCallback> {
         self.with_registry(Registry::credentials(authzid))
-    }
-}
-
-#[derive(Clone)]
-#[doc(hidden)]
-pub struct WantSorter {
-    mechanisms: Registry,
-}
-impl ConfigBuilder<WantSorter> {
-    /// Use the default mechanisms prioritizations
-    ///
-    /// This method is required to allow backwards-compatible expansion of the configuration builder
-    #[must_use]
-    pub fn with_defaults(self) -> ConfigBuilder<WantCallback> {
-        ConfigBuilder {
-            state: WantCallback {
-                mechanisms: self.state.mechanisms,
-                sorter: default_sorter,
-            },
-        }
     }
 }
 
@@ -162,7 +134,6 @@ impl ConfigBuilder<WantSorter> {
 #[doc(hidden)]
 pub struct WantCallback {
     mechanisms: Registry,
-    sorter: SorterFn,
 }
 impl ConfigBuilder<WantCallback> {
     /// Install a callback for querying properties
@@ -174,6 +145,6 @@ impl ConfigBuilder<WantCallback> {
         self,
         callback: CB,
     ) -> Result<Arc<SASLConfig>, SASLError> {
-        SASLConfig::new(callback, self.state.sorter, self.state.mechanisms)
+        SASLConfig::new(callback, self.state.mechanisms)
     }
 }