+++ src/mechanisms/oauthbearer/mechinfo.rs
@@ -1,6 +1,6 @@
 use crate::alloc::boxed::Box;
 use crate::mechname::Mechname;
-use crate::registry::{Mechanism, Side};
+use crate::registry::{Matches, Mechanism, Named, Side};
 
 #[cfg(feature = "registry_static")]
 use crate::registry::{distributed_slice, MECHANISMS};
@@ -14,7 +14,17 @@ use super::{client, server};
 pub static OAUTHBEARER: Mechanism = Mechanism {
     mechanism: Mechname::const_new(b"OAUTHBEARER"),
     priority: 300,
-    client: Some(|_sasl, _offered| Ok(Box::new(client::OAuthBearer::default()))),
+    client: Some(|| Ok(Box::new(client::OAuthBearer::default()))),
     server: Some(|_sasl| Ok(Box::new(server::OAuthBearer::default()))),
     first: Side::Client,
+
+    select: |_| Some(Matches::<Select>::name()),
+    offer: |_| true,
 };
+
+struct Select;
+impl Named for Select {
+    fn mech() -> &'static Mechanism {
+        &OAUTHBEARER
+    }
+}